+++ tests/pytest_helpers.py
@@ -7,6 +7,8 @@ from zipfile import ZipFile
 
 from skbuild import __version__ as skbuild_version
 
+from . import list_ancestors
+
 
 def check_wheel_content(wheel_archive, expected_distribution_name, expected_content, pure=False):
     """This function raises an AssertionError if the given wheel_archive
@@ -26,10 +28,20 @@ def check_wheel_content(wheel_archive, expected_distribution_name, expected_cont
     ]
 
     if parse_version(wheel.__version__) < parse_version('0.31.0'):
+        # These files were specified in the now-withdrawn PEP 426
+        # See https://github.com/pypa/wheel/issues/195
         expected_content += [
             '%s.dist-info/DESCRIPTION.rst' % expected_distribution_name,
             '%s.dist-info/metadata.json' % expected_distribution_name
         ]
+
+    if parse_version(wheel.__version__) > parse_version('0.33.1'):
+        # Include directory entries when building wheel
+        # See https://github.com/pypa/wheel/issues/287
+        directories = set()
+        for entry in expected_content:
+            directories = directories.union(list_ancestors(entry))
+        expected_content += directories
 
     if pure:
         assert wheel_archive.endswith('-none-any.whl')