+++ tests/test_include_exclude_data.py
@@ -1,14 +1,13 @@
 import glob
 import os
 import tarfile
-import wheel
 
-from pkg_resources import parse_version
 from zipfile import ZipFile
 
 from skbuild.utils import to_unix_path
 
 from . import project_setup_py_test
+from .pytest_helpers import check_wheel_content
 
 
 def check_whls(project_name):
@@ -17,10 +16,6 @@ def check_whls(project_name):
     assert not whls[0].endswith('-none-any.whl')
 
     expected_content = [
-        '%s.dist-info/top_level.txt' % project_name,
-        '%s.dist-info/WHEEL' % project_name,
-        '%s.dist-info/RECORD' % project_name,
-        '%s.dist-info/METADATA' % project_name,
         'hello/__init__.py',
         'hello/cmake_generated_module.py',
         'hello/data/subdata/hello_data1_include_from_manifest.txt',
@@ -37,15 +32,10 @@ def check_whls(project_name):
         'hello2/data2/subdata2/hello2_data2_include_from_manifest.txt',
         'hello2/hello2_include_from_manifest.txt',
     ]
-    if parse_version(wheel.__version__) < parse_version('0.31.0'):
-        expected_content += [
-            '%s.dist-info/DESCRIPTION.rst' % project_name,
-            '%s.dist-info/metadata.json' % project_name
-        ]
 
-    archive = ZipFile(whls[0])
-    member_list = archive.namelist()
-    assert sorted(expected_content) == sorted(member_list)
+    expected_distribution_name = project_name
+
+    check_wheel_content(whls[0], expected_distribution_name, expected_content)
 
 
 def check_sdist(proj, base=''):