+++ tests/test_manifest_in.py
@@ -9,12 +9,11 @@ Tries to build and test the `manifest-in` sample project.
 
 import glob
 import tarfile
-import wheel
 
-from pkg_resources import parse_version
 from zipfile import ZipFile
 
 from . import project_setup_py_test
+from .pytest_helpers import check_wheel_content
 
 
 @project_setup_py_test("manifest-in", ["sdist"], disable_languages_test=True)
@@ -51,21 +50,11 @@ def test_manifest_in_sdist():
 def test_manifest_in_wheel():
     whls = glob.glob('dist/*.whl')
     assert len(whls) == 1
-    assert whls[0].endswith('-none-any.whl')
 
     expected_content = [
-        'manifest_in-1.2.3.dist-info/top_level.txt',
-        'manifest_in-1.2.3.dist-info/WHEEL',
-        'manifest_in-1.2.3.dist-info/RECORD',
-        'manifest_in-1.2.3.dist-info/METADATA',
         'hello/__init__.py'
     ]
 
-    if parse_version(wheel.__version__) < parse_version('0.31.0'):
-        expected_content += [
-            'manifest_in-1.2.3.dist-info/DESCRIPTION.rst',
-            'manifest_in-1.2.3.dist-info/metadata.json'
-        ]
+    expected_distribution_name = 'manifest_in-1.2.3'
 
-    member_list = ZipFile(whls[0]).namelist()
-    assert sorted(expected_content) == sorted(member_list)
+    check_wheel_content(whls[0], expected_distribution_name, expected_content, pure=True)